refactor(condition): change new property to be type condition
