+++ src/Models/Properties/ElementProperties/OptionalIfProperty.cs
@@ -2,10 +2,6 @@
 {
     public partial class BaseProperty
     {
-        public string OptionalIfQuestionId { get; set; } = string.Empty;
-
-        public string OptionalIfValue { get; set; } = string.Empty;
-
-        public string OptionalIfNotValue { get; set; } = string.Empty;
+        public Condition OptionalIf { get; set; } = new();
     }
 }
